requirements.txt
tqdm
numpy
pandas
termcolor
gymnasium
pyyaml
tenacity

# cyberbattlesim @ git+https://github.com/MarcCote/CyberBattleSim.git@fix_gymnasium
cyberbattlesim @ git+https://github.com/microsoft/CyberBattleSim.git@main

datasets

#WikiSQL
#pysqlite3
#wikisql

# Code fixing
codebleu
tree-sitter<0.23.0
tree-sitter-c<0.23.0
tree-sitter-php<0.23.0
tree-sitter-python<0.23.0
tree-sitter-java<0.23.0
tree-sitter-cpp<0.23.0
tree-sitter-javascript<0.23.0
tree-sitter-ruby<0.23.0
tree-sitter-go<0.23.0
Levenshtein

# Data processing
scikit-learn

# baselines
openai
wandb